+++ scriptconfig/_ubelt_repr_extension.py
@@ -1,10 +1,12 @@
+from __future__ import annotations
 
-def _register_ubelt_repr_extensions():
+
+def _register_ubelt_repr_extensions() -> None:
     import ubelt as ub
     try:
-        _REPR_EXTENSIONS = ub.util_repr._REPR_EXTENSIONS
+        _REPR_EXTENSIONS = ub.util_repr._REPR_EXTENSIONS  # type: ignore[attr-defined]
     except AttributeError:
-        _REPR_EXTENSIONS = ub.util_format._FORMATTER_EXTENSIONS
+        _REPR_EXTENSIONS = ub.util_format._FORMATTER_EXTENSIONS  # type: ignore[attr-defined]
 
     def _register_scriptconfig_extensions():
         import scriptconfig as scfg